Finished Tank health, check everything, create ui

+++ Project/Assets/CodeBase/Logic/IArmor.cs
@@ -2,7 +2,7 @@
 {
     public interface IArmor
     {
-        int Armor { get; set; }
+        float Armor { get; set; }
 
     }
 }
